requirements.txt
tensorflow
numpy
opencv-python
pillow
streamlit
numpy
